requirements.txt
django==2.1.7
djangorestframework==3.9.2
psycopg2-binary==2.7.7
geopy==1.18.1
